+++ sort_photos.py
@@ -1,4 +1,5 @@
 #!/usr/bin/env python
+# formerly photoput
 # gets original date from EXIF information, and moves file to relative yyyy-MM-dd folder
 # Author: Jake Gustafson (poikilos)
 import sys
@@ -10,7 +11,7 @@ if len(sys.argv) < 2:
     exit(1)
 
 #process_files("/run/media/owner/sandisku32/DCIM/2017-10-29", 'move')
-process_files(sys.argv[1], 'move')
+results = process_files(sys.argv[1], 'move')
 
 print("unknown_type_count: " + str(results.get('unknown_type_count')))
 print("missing_meta_count: " + str(results.get('missing_meta_count')))